Dramatically reduce the number of subprocesses run
In nearly every case it is better to run subprocess.run() without the
burden of forking a bash shell too. All it takes is to turn strings into
arrays of strings and *remove* some use of formatting.

Aurman specifically, does not use *any* of the fancy shell
interpretation that sometimes makes forking a shell to fork the
commands, a worthwhile endeavor. So don't do so.

This is further cleanup after the removal of subprocesses for file
operations. Only changes where this cleanly applies have been done for
now... in a followup commit, subprocesses that handle arguments as
heavily-quoted strings will be handled.