remove possibility to do partial upgrades by using --aur or --repo
currently the use of --repo and --aur allows to do kind of partial upgrading, which can also be harmful. aur packages may depend on the current versions of packages in repos, use of --repo and --aur thus may lead to malfunctioning packages. aurman already disallows the use of "-Sy", this behavior is hence just consistent