Improve error on malformed known_package_places
The current traceback on a bad known_package_places file doesn’t make it
immediately obvious that it’s a local file’s JSON that has been garbled
rather than a response from AURweb or elsewhere that’s gone awry.

This commit adds a single message to the JSON decode error to note which
file the error is happening in and the user can then either delete the
file or try and fix it, depending on the state of the file in question.